+++ backend/models/BoardRequest.js
@@ -1,6 +1,6 @@
 const { DataTypes } = require('sequelize');
 const { sequelize } = require('../database/db.js');
-const { User } = require('./User.js');
+const User = require('./User');
 
 const BoardRequest = sequelize.define('BoardRequest', {
   request_id: {
