+++ backend/models/Comment.js
@@ -1,6 +1,6 @@
 const { DataTypes } = require('sequelize');
 const { sequelize } = require('../database/db.js');
-const { User } = require('./User.js');
+const User = require('./User');
 const { Post } = require('./Post.js');
 
 const Comment = sequelize.define('Comment', {
