+++ backend/models/ParticipatingChallenge.js
@@ -1,6 +1,6 @@
 const { sequelize, Sequelize } = require('../database/db');
 const { DataTypes } = Sequelize;
-const { User } = require('./User');
+const User = require('./User');
 const Challenge     = require('./Challenge');
 
 const ParticipatingChallenge = sequelize.define('ParticipatingChallenge', {
