+++ backend/models/participatingAttendance.js
